Support unique event payloads (#1)
Rather than attempting to combine all payload types into a single object, each GameAction is now associated with a unique struct which can define its own payload. All event structs are combined into a single EventUnion struct on the Event field of the Message. Making each field type a pointer effectively makes them optional types; this means on the frontend we need only send the field we want in our payload and on the backend we can simply update the switch statement to match on non-null cases.

Additionally, I have made all of the existing enums implement the String() interface and added a bit of polish to the log outputs, as well as resolved a couple breakages I ran into along the way.